Global Volume is integrated completely. There are few bugs in computation for SurfacePrediction

+++ KinectFusion/VirtualSensor_freiburg.h
@@ -62,7 +62,7 @@ public:
 
 		if ((unsigned int)m_currentIdx >= (unsigned int)m_filenameColorImages.size()) return false;
 
-		std::cout << "ProcessNextFrame [" << m_currentIdx << " | " << m_filenameColorImages.size() << "]" << std::endl;
+		std::cout << "ProcessNextFrame [" << (unsigned int)m_currentIdx << " | " << m_filenameColorImages.size() << "]" << std::endl;
 
 		FreeImageB rgbImage;
 		rgbImage.LoadImageFromFile(m_baseDir + m_filenameColorImages[m_currentIdx]);
